+++ index.js
@@ -37,4 +37,11 @@
 export {default} from './lib/router.js';
 export {default as graceful} from './lib/lifecycle.js';
 export {presets} from './lib/presets.js';
-export {defineGet, definePost, defineRoute} from './lib/define-route.js';
+export {
+  defineGet,
+  definePost,
+  defineRoute,
+  definePut,
+  definePatch,
+  defineDelete
+} from './lib/define-route.js';